test: breaking change

+++ pkg/utils/mailbox/mailbox.go
@@ -33,7 +33,7 @@ func NewSingle[T any]() *Mailbox[T] { return New[T](1) }
 
 // New creates a new mailbox instance. If name is non-empty, it must be unique and calling Start will launch
 // prometheus metric monitor that periodically reports mailbox load until Close() is called.
-func New[T any](capacity uint64) *Mailbox[T] {
+func New[T any](capacity uint32) *Mailbox[T] {
 	queueCap := capacity
 	if queueCap == 0 {
 		queueCap = 100
@@ -41,7 +41,7 @@ func New[T any](capacity uint64) *Mailbox[T] {
 	return &Mailbox[T]{
 		chNotify: make(chan struct{}, 1),
 		queue:    make([]T, 0, queueCap),
-		capacity: capacity,
+		capacity: uint64(capacity),
 	}
 }
